Setup a base test class instead
