+++ test/webpacker_test_helper.rb
@@ -20,5 +20,21 @@ module TestApp
   end
 end
 
+class Webpacker::Test < Minitest::Test
+  private
+    def reloaded_config
+      Webpacker.instance.instance_variable_set(:@config, nil)
+      Webpacker.config
+    end
+
+    def with_node_env(env)
+      original = ENV["NODE_ENV"]
+      ENV["NODE_ENV"] = env
+      yield
+    ensure
+      ENV["NODE_ENV"] = original
+    end
+end
+
 Rails.backtrace_cleaner.remove_silencers!
 TestApp::Application.initialize!
